requirements.txt.txt
asgiref                       
attrs                         
certifi                       
cffi                          
charset-normalizer            
cryptography                  
defusedxml
dj-rest-auth
Django
django-allaut  
djangorestframework 
djangorestframework-simplejwt    
drf-spectacular  
idna
inflection
jsonschema
jsonschema-specifications
oauthlib
pip
pycparser
PyJWT
python3-openid
PyYAML
referencing
requests
requests-oauthlib
rpds-py
setuptools
sqlparse
tzdata
uritemplate
urllib3